// AIRMAM5/Scripts/Models/Interface/Subject/KeyFrameDataModel.ts
/**關鍵影格資訊 */
export interface KeyFrameDataModel {
    /**檔案編號 */
    fsFILE_NO: string;
    /**標題 */
    Title: string;
    /**描述 */
    Description: string;
    /**是否為代表圖 */
    IsHeadFrame: boolean;
    /**檔案路徑Path */
    FilePath: string;
    /**時間 */
    Time: string;
    /**截圖URL */
    ImageUrl: string;
    /**檔案資訊 */
    FileInfo: string;
}